api/requirements.txt
Flask==3.0.0
SPARQLWrapper==2.0.0
python-dotenv==1.0.1
flask-cors==4.0.0